# GH-99298: Clean up attribute specializations